Experimentally special-case “Cannot recover” err
Many real-world HTML documents contain markup cases that cannot be parsed
in a purely-streaming parser, so when the parser is used in the context of
the HTML checker, this “Cannot recover” exception gets thrown so often,
with an accompanying stack trace output to the logs, that it basically just
generates a huge amount of noise in the logs.

This change creates a special exception for this case, so that we can catch
it in the HTML checker code (and prevent the stack trace for getting logged
every single time).