Experimentally special-case “Changing encoding” err
Many real-world HTML documents on the Web contain `meta` elements deep
inside the documents which trigger the parser to emit a “Changing encoding
at this point would need non-streamable behavior.” error. So when the
parser is used in the context of the HTML checker, that error gets output
to the log so often, with an accompanying stack trace, that it basically
just generates a unwelcome noise in the logs.

This change creates a special exception for this case, so that we can catch
it in the HTML checker code (and prevent the stack trace from getting
logged every single time).