Enable control for tokenizer buffer size
This change adds an optional bufferSize parameter to the tokenize(), parse(),
and parseFragment() methods of sax.HtmlParser instances. That bufferSize
parameter controls the size of the buffer which gets fed to the tokenizer.

The control provided by that parameter allows the tokenizer buffer to be
set, for example, to "1" — and that is particularly useful for emulating
the behavior of the Firefox HTML parser, which feeds the tokenizer one
single code unit at a time.

Otherwise, without this change, the tokenizer buffer size for HtmlParser
instances is hardcoded to 2048.