Set dontSwallowBom for testharness
This change causes the dontSwallowBom() method to be called for the
driver instance used in TokenizerTester. That causes any BOM at the
beginning of test input to not be swallowed but instead preserved.

Otherwise, without this change, the following html5lib-tests case fails:

* https://github.com/html5lib/html5lib-tests/blob/master/tokenizer/domjs.test#L197

Relates to https://github.com/validator/htmlparser/issues/35